Fix redirects from error page
Prior to this fix, clicking on the header from a 404 error page
WOULD change the path to '/', but would NOT actually change the
content of the page from "Resource not found" to the typical home
page.

This was the case because the "isValidState" member of <AppBody />'s
state would carry over through the path change, and when it
re-rendered, would cause the error page to show up again.

This fix causes <AppBody /> to run through its snippet loading
logic whenever it is
  1) mounted, or
  2) receives new props, and the router's path has changed.