Store user data in redux, not cookies
Up to this point, we have stored user information (such as their
username, avatar URL, orgs, etc...) in cookies. This has led to
problems with stale data.

This commit moves such information into the redux state, rather
than the cookies. Cookies are only used when necessary (for now,
this means storing the token, signInRedirect, and signInState).

<CodesplainAppBar /> is now responsible for loading the token from
the cookies, and using it to obtain information about the user and
storing it in redux, for use by the other components. This means
that every time the page is loaded (including a refresh), the
user's data is updated, preventing stale data problems.